move azure-functions test code out of the 'modules/' subdir to match to pattern in lib/...
