+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/.vscode/extensions.json
