+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFn1/function.json
