+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnContextDone/function.json
