+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnContextDone/index.js
