+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnError/function.json
