+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnError/index.js
