+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnReturnContext/function.json
