+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnReturnContext/index.js
