+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnReturnObject/function.json
