+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnReturnObject/index.js
