+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnReturnString/function.json
