+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/HttpFnReturnString/index.js
