+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/host.json
