+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/local.settings.json
