+++ test/instrumentation/azure-functions/fixtures/AJsAzureFnApp/package.json
